omr-sheet-recognition-main/omr-sheet-recognition-main/requirements.txt
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.24.0
opencv-python==4.8.0.74
ipywidgets==8.0.7
imutils==0.5.4
pdf2image==1.16.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
scipy==1.11.1
pytesseract==0.3.10
black==23.7.0
ruff==0.0.280
rope==1.9.0
numpy~=1.25.2
pandas~=2.0.3
Pillow~=10.0.0